refactor: some renaming
