+++ .github/workflows/release-please.yml
@@ -2,17 +2,13 @@ name: Release Please
 
 on:
   push:
-    tags:
-      - "v*"
-
-permissions:
-  contents: write
-  id-token: write
-  pull-requests: write
+    branches: [main]
 
 jobs:
   release-please:
     runs-on: ubuntu-latest
+    permissions:
+      pull-requests: write
     steps:
       - uses: google-github-actions/release-please-action@v3
         id: release
@@ -26,6 +22,9 @@ jobs:
     runs-on: ubuntu-latest
     needs: release-please
     if: ${{ needs.release-please.outputs.release_created }}
+    permissions:
+      contents: write
+      id-token: write
     steps:
       - uses: actions/checkout@v4
 
@@ -37,8 +36,8 @@ jobs:
           cache: "npm"
           registry-url: "https://registry.npmjs.org"
 
-      - name: Build NPM package
-        run: deno task npm
+      - name: Build
+        run: deno task build
 
       - name: Publish NPM package
         working-directory: dist/npm
@@ -46,10 +45,7 @@ jobs:
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
 
-      - name: Build tar files
+      - name: Upload the palette files
         run: |
-          deno task thirdparty
           tar czf catppuccin.tar.gz --directory=dist/palettes/ .
-
-      - name: Attach files to release
-        run: gh release upload ${{ needs.release-please.outputs.tag_name }} catppuccin.tar --clobber
+          gh release upload ${{ needs.release-please.outputs.tag_name }} catppuccin.tar --clobber
