+++ .gitignore
@@ -1,62 +1,2 @@
-# General
-!*.keep
-
-# building
-out/
-.env
-*~
-/bin
-.nfs*
-$RECYCLE.BIN/
-build
-build_output/*
 dist/
-*.exe
-*.out
-*.app
-*.o
-*.so
-
-# logs
-*.log
-*.bak
-*.errors
-*.stackdump
-crash-*
-
-# tmp files
-*.temp
-*.tmp
-/tmp/*
-/tmp
-.cache
-*.pid
-*.seed
-*.pid.lock
-
-# git
-*.patch
-*.diff
-
-# IDEs/DEs
-.idea/
-.vscode/*
-*.sublime-*
-.history
-tags
-
-# dev platforms/package managers
-.npm
-.yarn/*
 node_modules/
-jspm_packages/
-
-# OSs
-*.linux
-.DS_Store
-.fuse_hidden*
-.Trash-*
-.directory
-[Dd]esktop.ini
-.Trashes
-*.lnk
