+++ scripts/builders/mod.ts
@@ -1,9 +1,9 @@
-export { compileCss } from "./css.ts";
-export { compileLess } from "./less.ts";
-export { compileScss } from "./scss.ts";
+export { compileCss } from "./npm/css.ts";
+export { compileLess } from "./npm/less.ts";
+export { compileScss } from "./npm/scss.ts";
 
-export { generateAse } from "./thirdparty/ase.ts";
-export { generateGimp } from "./thirdparty/gimp.ts";
-export { generatePng } from "./thirdparty/png.ts";
-export { generateProcreate } from "./thirdparty/procreate.ts";
-export { generateSip } from "./thirdparty/sip.ts";
+export { generateAse } from "./palettes/ase.ts";
+export { generateGimp } from "./palettes/gimp.ts";
+export { generatePng } from "./palettes/png.ts";
+export { generateProcreate } from "./palettes/procreate.ts";
+export { generateSip } from "./palettes/sip.ts";
