+++ scripts/builders/palettes/gimp.ts
