+++ scripts/builders/palettes/png.ts
